Moved AB3 Person classes into its own packages
Created MeetingEntry packages in Parser and Command
