+++ src/main/java/seedu/address/logic/commands/meetingentry/AddCommand.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.meetingentry;
+
+public class AddCommand {
+}
