+++ src/main/java/seedu/address/logic/parser/meetingentry/AddCommandParser.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.parser.meetingentry;
+
+public class AddCommandParser {
+}
